move view/webContents helpers to lib/web-contents
applyViewZoomLimits, openViewDevToolsInDev, and
broadcastFoundInPageResults operate on a WebContentsView (and a
WebContents target in the last case), not on BrowserWindow shells.
Split them out of lib/window.ts and ipc.ts into a new
lib/web-contents.ts so lib/window stays focused on BrowserWindow
plumbing and ipc.ts on IPC registration.
